refactor: 폴더명 team_member -> teamMember로 변경
