+++ src/main/java/PNUMEAT/Backend/domain/team/service/TeamService.java
@@ -5,8 +5,8 @@ import PNUMEAT.Backend.domain.team.dto.request.TeamRequest;
 import PNUMEAT.Backend.domain.team.entity.Team;
 import PNUMEAT.Backend.domain.team.enums.Topic;
 import PNUMEAT.Backend.domain.team.repository.TeamRepository;
-import PNUMEAT.Backend.domain.team_member.entity.TeamMember;
-import PNUMEAT.Backend.domain.team_member.repository.TeamMemberRepository;
+import PNUMEAT.Backend.domain.teamMember.entity.TeamMember;
+import PNUMEAT.Backend.domain.teamMember.repository.TeamMemberRepository;
 import PNUMEAT.Backend.global.images.ImageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
